\name{flowQ-package}
\alias{flowQ-package}
\alias{flowQ}
\docType{package}
\title{
  Quality control for flow cytometry
}
\description{
  Functions and methods for quality control and QA of flow cytometry
  data. This package heavily depends on the flowCore package.  }

\details{ \tabular{ll}{
    Package: \tab flowQ
    \cr Type: \tab Package
    \cr Version: \tab 0.2.1
    \cr Date: \tab 2006-11-16
    \cr License: \tab Artistic\cr }

Quality control is an important aspect when dealing with large amounts
of complex high-throughput data. This package comprises functionality
for efficient QA of flow cytometry data.
}
\author{
  Maintainer: Florian Hahne <f.hahne@dkfz.de>
  Authors: R. Gentleman, F. Hahne, J. Kettman, N. Le Meur, M. Tang
}
\references{
references go here
}

\keyword{ package }
\seealso{
\code{\link[flowCore:flowCore-package]{flowCore}}
}
\examples{
## examples go here
}
