\name{outlier-class}
\docType{class}
\alias{outlier-class}

\title{Virtual Class "outlier"}
\description{A class to represent outlier tests}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
	 \describe{
    \item{\code{test}:}{Object of class \code{"character"} ~~ }
    \item{\code{parameters}:}{Object of class \code{"ANY"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "outlier" in the signature.
}
\references{ ~put references to the literature/web site here ~ }

\examples{
## showClass("outlier")
}
\keyword{classes}

