\name{outlierResult-class}
\docType{class}
\alias{outlierResult-class}

\title{Class "outlierResult"}
\description{A class to hold the results of an outlier test}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("outlierResult", ...)}.
	 ~~ describe objects here ~~ 
}
\section{Slots}{
	 \describe{
    \item{\code{frameId}:}{Object of class \code{"character"} ~~ }
    \item{\code{filterDetails}:}{Object of class \code{"list"} ~~ }
    \item{\code{test}:}{Object of class \code{"character"} ~~ }
    \item{\code{parameters}:}{Object of class \code{"ANY"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{outlier}"}, directly.
}
\section{Methods}{
No methods defined with class "outlierResult" in the signature.
}
\references{ ~put references to the literature/web site here ~ }
\examples{
#showClass("outlierResult")
}
\keyword{classes}
