\name{qaProcess.2DStatsPlot}
\alias{qaProcess.2DStatsPlot}

\title{  Creates a QA process for 2D summary statistic comparisons across
       multiple samples }
\description{
   This function takes a list of \code{\link[flowCore]{flowSet}} as input and
  creates all necessary output for a '2DStatsPlot' type QA
  process. Objects created by this function can be laid out as HTML
  using \code{\link{writeQAReport}}.
}
\usage{
 
 qaProcess.2DStatsPlot(flowList,dyes=c("FSC-A","SSC-A"),
			  outdir="QAReport",outBound=0.25,func=mean,
			  det.dimensions=c(400,400),pdf=FALSE,...)
}

\arguments{
  
  \item{flowList}{ A list of \code{\link[flowCore]{flowSet}} }
  \item{dyes}{Flow parameters to be compared across multiple samples}
  \item{outdir}{ The directory to which the graphical output is to be
    saved. If multiple QA processes are to be combined, make sure to use
    the same directory every time. }
  \item{outBound}{ A numeric value between 0 and 1 which is used as a threshold
  by the outlier detection algorithm}
 \item{func}{Summary statistic function to be applied to each parameter in a flow Frame
  (mean,median etc)
 }  
 \item{det.dimensions}{the dimensions of each image generated by the QA process.}
 \item{pdf}{boolen value determinining if pdf files or jpeg images will be 
	    produced by the QA process} 
  \item{...}{Further arguments.}
 }
 \details{
   
  QA processes of type '2DStatsPlot' detects differences in the value of 
  a summary statistic such as the mean, median etc across multiple 
  samples.
  
  For more details on how to layout \code{\link{qaProcess}} objects to
  HTML, see \code{\link{writeQAReport}} and \code{\link{qaReport}}.
  
}
\value{
  
  An object of class \code{\link{qaProcess}}.
  
}

\author{ Nishant Gopalakrishnan }


\seealso{
  \code{\link{writeQAReport}}, \code{\link{qaReport}},
  \code{\link{qaProcess}}, \code{\link{qaProcess.marginevents}},
  \code{\link{qaProcess.DensityPlot}}, \code{\link{qaProcess.ECDFPlot}}

}


\examples{
data(qData)
dest <- tempdir()
par<-c("FSC-A","SSC-A","CD4","CD8")
resMean <- qaProcess.2DStatsPlot(qData,dyes=par,outdir=dest,func=mean,outBound=0.28,pdf=TRUE)
resMean
%summaryUrl<-writeQAReport(qData[[1]], list(resMean), outdir=dest,pdf=TRUE)
%browseURL(summaryUrl)

}

\keyword{ dynamic }


 
