\name{qaProcess.BoundaryPlot}
\alias{qaProcess.BoundaryPlot}

\title{  Creates a QA process for comparison of the percentage of boundary 
events for the specified parameters across multiple samples}
\description{
   This function takes a list of \code{\link[flowCore:flowSet-class]{flowSet}}
   as input and 
  creates all necessary output for a 'BoundaryPlot' type QA
  process. Objects created by this function can be laid out as HTML
  using \code{\link[flowQ]{writeQAReport}}.
}
\usage{
 
 qaProcess.BoundaryPlot(flowList,dyes=NULL,
			  outdir="QAReport",cutoff=3,sum.dimensions=NULL,
			  det.dimensions=NULL,pdf=TRUE,name="Boundary",
			  side="both",...)
}

\arguments{
  
  \item{flowList}{ A list of \code{\link[flowCore:flowSet-class]{flowSet}} }
  \item{dyes}{Flow parameters to be compared across multiple samples. 
     If set to NULL, all parameters that are duplicated across the list of 
     \code{\link[flowCore:flowSet-class]{flowSet}} are identified and compared.}
  \item{outdir}{ The directory to which the graphical output is to be
    saved. If multiple QA processes are to be combined, make sure to use
    the same directory every time. }
  \item{cutoff}{ A numeric value for the percentage of boundary events that is 
    used by the  process to identify samples that failed the QA process}
   \item{sum.dimensions}{The dimensions of summary plot generated in inches, default
 is NULL where the dimensions are automatically selected}
 \item{det.dimensions}{the dimensions of each sub image generated by the QA
  process in inches, default is NULL where the dimensions are automatically
  selected.} 
 \item{pdf}{boolen value determinining if pdf files or jpeg images will
    be produced by the QA process}  
  \item{name}{The name of the process used for the headings in the HTML
    output.
  }
  \item{side}{Object of class \code{"character"}. The margin
       on which to evaluate the filter. Either \code{upper} for the
       upper margin or \code{lower} for the lower margin or \code{both}
       for both margins. }
  \item{...}{Further arguments.}
 }
 \details{
   
  QA processes of type 'BoundaryPlot'  helps identify samples that have a large 
number of boundary events that must be removed by further gating before 
detailed analysis of the data.
  
  For more details on how to layout
  \code{\link[flowQ:qaProcess-class]{qaProcess}} objects to 
  HTML, see \code{\link[flowQ]{writeQAReport}} and
  \code{\link[flowQ]{qaReport}}. 
  
}
\value{
  
  An object of class \code{\link[flowQ:qaProcess-class]{qaProcess}}.
  
}

\author{ Nishant Gopalakrishnan }


\seealso{
  \code{\link[flowQ]{writeQAReport}},
  \code{\link[flowQ]{qaReport}},
  \code{\link[flowQ:qaProcess-class]{qaProcess}},
  \code{\link[flowQ]{qaProcess.DensityPlot}},
  \code{\link[flowQ]{qaProcess.ECDFPlot}}

}


\examples{
data(qData)
dest <- tempdir()
resBoundary <- qaProcess.BoundaryPlot(qData, dyes=c("FSC-A","CD3"), 
                                      outdir=dest, cutoff=3, pdf=TRUE)
resBoundary
%writeQAReport(qData[[1]], list(resBoundary), outdir=dest,pdf=TRUE)
}

\keyword{ dynamic }


 
