\name{qaProcess.DensityPlot}
\alias{qaProcess.DensityPlot}

\title{ Creates a QA process for comparisons of density plots across 
        multiple samples }
\description{
   This function takes a list of \code{\link[flowCore]{flowSet}} as input and
  creates all necessary outputs for a 'DensityPlot' type QA
  process. Objects created by this function can be laid out as HTML
  using \code{\link{writeQAReport}}.
}
\usage{
  qaProcess.DensityPlot(flowList,dyes=NULL,outdir="QAReport",
		       alpha=0.05,absolute.value=NULL,det.dimensions=c(7,7),
		     pdf=TRUE,name="Density",...)
}

\arguments{
  \item{flowList}{ A list of \code{\link[flowCore]{flowSet}} }
  \item{dyes}{Flow parameters to be compared across multiple samples. 
If set to NULL, all parameters that are duplicated across the list of 
\code{\link[flowCore]{flowSet}} are identified and compared.}
  \item{outdir}{ The directory to which the graphical output is to be
    saved. If multiple QA processes are to be combined, make sure to use
    the same directory every time. }
  \item{alpha}{Outlier mislabeling rate}
  \item{absolute.value}{If provided, the absolute.value is used for the 
  outlier detection and the alpha argument is ignored. defualt is NULL}
 \item{det.dimensions}{the dimensions of each image generated by the QA process.}
 \item{pdf}{boolen value determinining if pdf files or jpeg images will be 
	    produced by the QA process} 
 \item{name}{The name of the process used for the headings in the HTML
    output.}
  \item{...}{Further arguments.}
}
 \details{
   
  QA processes of type 'DensityPlot' detects differences in the density plots
  across multiple samples. A summary measure based on the sum of the pairwise KL
  Distances of density plots is used as a parameter to flag patient panels that  
  are potential outliers.
  
  For more details on how to layout \code{\link{qaProcess}} objects to
  HTML, see \code{\link{writeQAReport}} and \code{\link{qaReport}}.
  
}
\value{
  
  An object of class \code{\link{qaProcess}}.  
}

\author{ Nishant Gopalakrishnan }


\seealso{
  \code{\link{writeQAReport}}, \code{\link{qaReport}},
  \code{\link{qaProcess}}, \code{\link{qaProcess.marginevents}},
  \code{\link{qaProcess.ECDFPlot}}, \code{\link{qaProcess.2DStatsPlot}}
}


\examples{
data(qData)
dest <- tempdir()
resDensityFSC <- qaProcess.DensityPlot(qData,dyes=c("FSC-A","SSC-A"),outdir=dest,alpha=0.2,pdf=TRUE)
%densityUrl<-writeQAReport(qData[[1]], list(resDensity), outdir=dest)
%browseURL(densityUrl)

}

\keyword{ dynamic }


 
