\name{qaProcess.KLDistPlot}
\alias{qaProcess.KLDistPlot}

\title{  Creates a QA process for comparisons of KL Distances across
  multiple samples } 
\description{
   This function takes a list of
  \code{\link[flowCore:flowSet-class]{flowSet}} as input and 
  creates all necessary outputs for an 'KLDistPlot' type of QA
  process. Objects created by this function can be laid out as HTML
  using \code{\link[flowQ]{writeQAReport}}.
}
\usage{
 
 qaProcess.KLDistPlot(flowList,dyes=NULL,outdir="QAReport",
			alpha=0.05, absolute.value=NULL,sum.dimensions=NULL,
			det.dimensions=NULL, pdf=TRUE,
			name="KLDist",...)
}

\arguments{
  
  \item{flowList}{ A list of \code{\link[flowCore:flowSet-class]{flowSet}} }
 \item{dyes}{Flow parameters to be compared across multiple samples. 
   If set to NULL, all parameters that are duplicated across the list of 
    \code{\link[flowCore:flowSet-class]{flowSet}} are identified and compared.}
  \item{outdir}{ The directory to which the graphical output is to be
    saved. If multiple QA processes are to be combined, make sure to use
    the same directory every time. }
 \item{alpha}{Outlier mislabeling rate}
 \item{absolute.value}{If provided, the absolute.value is used for the 
  outlier detection and the alpha argument is ignored. defualt is NULL}
   \item{sum.dimensions}{The dimensions of summary plot generated in inches, default
 is NULL where the dimensions are automatically selected}
 \item{det.dimensions}{the dimensions of each sub image generated by the QA
  process in inches, default is NULL where the dimensions are automatically
  selected.} 
 \item{pdf}{boolen value determinining if pdf files or jpeg images will be 
	    produced by the QA process} 
 \item{name}{The name of the process used for the headings in the HTML
    output.}
  \item{...}{Further arguments.}

 }
 \details{
   
  QA processes of type 'KLDistPlot' detects differences in the pairwise
  KL Distances of the flow parameters across multiple samples.
  
  For more details on how to layout
  \code{\link[flowQ:qaProcess-class]{qaProcess}} objects to 
  HTML, see \code{\link[flowQ]{writeQAReport}} and
  \code{\link[flowQ]{qaReport}}. 
  
}
\value{
  
  An object of class \code{\link[flowQ:qaProcess-class]{qaProcess}}.
  
}

\author{ Nishant Gopalakrishnan }


\seealso{
    \code{\link[flowQ]{writeQAReport}},
  \code{\link[flowQ]{qaReport}},
  \code{\link[flowQ:qaProcess-class]{qaProcess}},
  \code{\link[flowQ]{qaProcess.DensityPlot}}
  \code{\link[flowQ]{qaProcess.2DStatsPlot}}
  \code{\link[flowQ]{qaProcess.BoundaryPlot}}

}


\examples{
data(qData)
dest <- tempdir()
resKLdist <- qaProcess.KLDistPlot(qData, dyes=c("SSC-A","FSC-A"),
               outdir=dest, alpha=0.05, pdf=TRUE)
resKLdist
%writeQAReport(qData[[1]], list(resKLdist), outdir=dest)

}

\keyword{ dynamic }


 
