\name{qaProcess.summaryOutLiers}
\alias{qaProcess.summaryOutLiers}

\title{  Creates a QA process for 2D summary statistic comparisons across
       multiple samples }
\description{
   This function takes a list of \code{\link[flowCore]{flowSet}} as input and
  creates all necessary output for a 'summaryOutLiers' type QA
  process. Objects created by this function can be laid out as HTML
  using \code{\link{writeQAReport}}.
}
\usage{
 
 qaProcess.summaryOutLiers(flowList,channels=c("FSC-A","SSC-A"),
			  outdir="QAReport",cutoff=4,func=median,
			  det.dimensions=c(300,300),pdf=FALSE,...)
}

\arguments{
  
  \item{flowList}{ A list of \code{\link[flowCore]{flowSet}} }
  \item{channels}{Flow parameters to be compared across multiple samples}
  \item{outdir}{ The directory to which the graphical output is to be
    saved. If multiple QA processes are to be combined, make sure to use
    the same directory every time. }
  \item{cutoff}{ The threshold at which the QA process is considered to
    fail.}
 \item{func}{Summary statistic function to be applied to each parameter in a flow Frame }  
 \item{det.dimensions}{the dimensions of each image generated by the QA process.}
 \item{pdf}{boolen value determinining if pdf files or jpeg images will be 
	    produced by the QA process} 
  \item{...}{Further arguments.}
 }
 \details{
   
  QA processes of type 'summaryOutLiers' detects differences in the value of 
  a summary statistic such as the mean, median, IQR etc across multiple 
  samples.
  
  For more details on how to layout \code{\link{qaProcess}} objects to
  HTML, see \code{\link{writeQAReport}} and \code{\link{qaReport}}.
  
}
\value{
  
  An object of class \code{\link{qaProcess}}.
  
}

\author{ Nishant Gopalakrishnan }


\seealso{
  \code{\link{writeQAReport}}, \code{\link{qaReport}},
  \code{\link{qaProcess}}, \code{\link{qaProcess.marginevents}},
  \code{\link{qaProcess.densityOutLiers}}, \code{\link{qaProcess.ecdfOutLiers}}

}


\examples{

\dontrun{

}

}

\keyword{ dynamic }


 
