\name{validProcess}
\alias{validProcess}

\title{ Validate a QAProcess object }
\description{
  Check for integrety and existence of files specified in a \code{\link{qaProcess}} object
}
\usage{
validProcess(object)
}

\arguments{
  \item{object}{ A \code{\link{qaProcess}} object }
}

\value{
 The function is called for it's side effects
}

\author{ Florian Hahne }

\seealso{ \code{\link{qaProcess.marginevents}}, \code{\link{qaReport}},
  \code{\link{qaProcess}}, \code{\link{qaProcess.timeline}},
  \code{\link{writeQAReport}}
}
% \examples{

% }

\keyword{ misc }
