\name{qaProcess.2DStatsPlot}
\alias{qaProcess.2DStatsPlot}

\title{  Creates a QA process for 2D summary statistic comparisons across
       multiple samples }
\description{
   This function takes a list of
       \code{\link[flowCore:flowSet-class]{flowSet}} as input and 
  creates all necessary output for a '2DStatsPlot' type QA
  process. Objects created by this function can be laid out as HTML
  using \code{\link[flowQ]{writeQAReport}}.
}
\usage{
 
 qaProcess.2DStatsPlot(flowList,dyes=c("FSC-A","SSC-A"),
			  outdir="QAReport",thresh=0.25,func=mean,
			  sum.dimensions=NULL,det.dimensions= NULL,pdf=TRUE,
			  name="2DStats",...)
}

\arguments{
  
  \item{flowList}{ A list of \code{\link[flowCore:flowSet-class]{flowSet}} }
  \item{dyes}{Flow parameters to be compared across multiple samples}
  \item{outdir}{ The directory to which the graphical output is to be
    saved. If multiple QA processes are to be combined, make sure to use
    the same directory every time. }
  \item{thresh}{ A numeric value between 0 and 1 which is used as a threshold
  by the outlier detection algorithm}
 \item{func}{Summary statistic function to be applied to each parameter
  in a \code{flowFrame} (mean,median etc)
 }  
 \item{sum.dimensions}{The dimensions of summary plot generated in inches, default
 is NULL where the dimensions are automatically selected}
 \item{det.dimensions}{the dimensions of each sub image generated by the QA
  process in inches, default is NULL where the dimensions are automatically
  selected.} 
 \item{pdf}{boolen value determinining if pdf files or jpeg images will be 
	    produced by the QA process} 
 \item{name}{The name of the process used for the headings in the HTML
    output.}
  \item{...}{Further arguments.}
 }
 \details{
   
  QA processes of type '2DStatsPlot' detects differences in the value of 
  a summary statistic such as the mean, median etc across multiple 
  samples.
  
  For more details on how to layout
  \code{\link[flowQ:qaProcess-class]{qaProcess}} objects to 
  HTML, see \code{\link[flowQ]{writeQAReport}} and
  \code{\link[flowQ]{qaReport}}. 
  
}
\value{
  
  An object of class \code{\link[flowQ:qaProcess-class]{qaProcess}}.
  
}

\author{ Nishant Gopalakrishnan }


\seealso{
  \code{\link[flowQ]{writeQAReport}},
  \code{\link[flowQ]{qaReport}},
  \code{\link[flowQ:qaProcess-class]{qaProcess}},
  \code{\link[flowQ]{qaProcess.marginevents}},
  \code{\link[flowQ]{qaProcess.DensityPlot}},
  \code{\link[flowQ]{qaProcess.ECDFPlot}}

}


\examples{
data(qData)
dest <- tempdir()
par<-c("FSC-A","SSC-A","CD4","CD8")
resMean <- qaProcess.2DStatsPlot(qData,dyes=par,
                outdir=dest, func=mean, thresh=0.28,pdf=TRUE) 
resMean
%summaryUrl<-writeQAReport(qData[[1]], list(resMean), outdir=dest,pdf=TRUE)
%browseURL(summaryUrl)

}

\keyword{ dynamic }


 
