\name{qaProcess.BoundaryPlot}
\alias{qaProcess.BoundaryPlot}

\title{  Creates a QA process for comparison of the percentage of boundary 
events for the specified parameters across multiple samples}
\description{
   This function takes a list of \code{\link[flowCore]{flowSet}} as input and
  creates all necessary output for a 'BoundaryPlot' type QA
  process. Objects created by this function can be laid out as HTML
  using \code{\link{writeQAReport}}.
}
\usage{
 
 qaProcess.BoundaryPlot(flowList,dyes=NULL,
			  outdir="QAReport",cutoff=3,
			  det.dimensions=c(400,400),pdf=FALSE,...)
}

\arguments{
  
  \item{flowList}{ A list of \code{\link[flowCore]{flowSet}} }
  \item{dyes}{Flow parameters to be compared across multiple samples. 
If set to NULL, all parameters that are duplicated across the list of 
\code{\link[flowCore]{flowSet}} are identified and compared.
	     }
  \item{outdir}{ The directory to which the graphical output is to be
    saved. If multiple QA processes are to be combined, make sure to use
    the same directory every time. }
  \item{cutoff}{ A numeric value for the percentage of boundary events that is 
used by the  process to identify samples that failed the QA process
}
 \item{det.dimensions}{the dimensions of each image generated by the QA process.}
 \item{pdf}{boolen value determinining if pdf files or jpeg images will be 
	    produced by the QA process} 
  \item{...}{Further arguments.}
 }
 \details{
   
  QA processes of type 'BoundaryPlot'  helps identify samples that have a large 
number of boundary events that must be removed by further gating before 
detailed analysis of the data.
  
  For more details on how to layout \code{\link{qaProcess}} objects to
  HTML, see \code{\link{writeQAReport}} and \code{\link{qaReport}}.
  
}
\value{
  
  An object of class \code{\link{qaProcess}}.
  
}

\author{ Nishant Gopalakrishnan }


\seealso{
  \code{\link{writeQAReport}}, \code{\link{qaReport}},
  \code{\link{qaProcess}},\code{\link{qaProcess.DensityPlot}}, \code{\link{qaProcess.ECDFPlot}}

}


\examples{

\dontrun{

}

}

\keyword{ dynamic }


 
