\name{qaProcess.KLDistPlot}
\alias{qaProcess.KLDistPlot}

\title{  Creates a QA process for comparisons of KL Distances across multiple samples }
\description{
   This function takes a list of \code{\link[flowCore]{flowSet}} as input and
  creates all necessary outputs for an 'KLDistPlot' type of QA
  process. Objects created by this function can be laid out as HTML
  using \code{\link{writeQAReport}}.
}
\usage{
 
 qaProcess.KLDistPlot(flowList,dyes=NULL,outdir="QAReport",
			alpha=0.05,det.dimensions=c(400,400),pdf=FALSE,...)
}

\arguments{
  
  \item{flowList}{ A list of \code{\link[flowCore]{flowSet}} }
 \item{dyes}{Flow parameters to be compared across multiple samples. 
   If set to NULL, all parameters that are duplicated across the list of 
    \code{\link[flowCore]{flowSet}} are identified and compared.}
  \item{outdir}{ The directory to which the graphical output is to be
    saved. If multiple QA processes are to be combined, make sure to use
    the same directory every time. }
 \item{alpha}{Outlier mislabeling rate}
 \item{det.dimensions}{the dimensions of each image generated by the QA process.}
 \item{pdf}{boolen value determinining if pdf files or jpeg images will be 
	    produced by the QA process} 
  \item{...}{Further arguments.}

 }
 \details{
   
  QA processes of type 'KLDistPlot' detects differences in the pairwise KL Distances
  of the flow parameters across multiple samples.
  
  For more details on how to layout \code{\link{qaProcess}} objects to
  HTML, see \code{\link{writeQAReport}} and \code{\link{qaReport}}.
  
}
\value{
  
  An object of class \code{\link{qaProcess}}.
  
}

\author{ Nishant Gopalakrishnan }


\seealso{
  \code{\link{writeQAReport}}, \code{\link{qaReport}},
  \code{\link{qaProcess}}, \code{\link{qaProcess.BoundaryPlot}},
  \code{\link{qaProcess.DensityPlot}}, \code{\link{qaProcess.2DStatsPlot}}
}


\examples{
data(qData)
dest <- tempdir()
resKLdist <- qaProcess.KLDistPlot(qData,dyes=c("SSC-A","FSC-A"),outdir=dest,alpha=0.05,pdf=TRUE)
resKLdist
%writeQAReport(qData[[1]], list(resKLdist), outdir=dest)

}

\keyword{ dynamic }


 
