\name{validProcess}
\alias{validProcess}

\title{ Validate a QAProcess object }
\description{
  Check for integrety and existence of files specified in a
  \code{\link[flowQ:qaProcess-class]{qaProcess}} object.
  
}
\usage{
validProcess(object)
}

\arguments{
  \item{object}{ A \code{\link[flowQ:qaProcess-class]{qaProcess}} object }
}

\value{
 The function is called for it's side effects.
}

\author{ Florian Hahne }

\seealso{ \code{\link[flowQ]{qaProcess.marginevents}},
  \code{\link[flowQ]{qaReport}},
  \code{\link[flowQ:qaProcess-class]{qaProcess}},
  \code{\link[flowQ]{qaProcess.timeline}},
  \code{\link[flowQ]{writeQAReport}}
}
% \examples{

% }

\keyword{ misc }
